Move cast to int function into cleaning utils
