+++ tests/test_file_data.py
@@ -248,68 +248,6 @@ class ASCWDSWorkplaceData:
         ),
     ]
 
-    cast_to_int_rows = [
-        (
-            "loc 1",
-            "20",
-            "18",
-        ),
-    ]
-
-    cast_to_int_errors_rows = [
-        (
-            "loc 1",
-            "20",
-            "18",
-        ),
-        (
-            "loc 2",
-            "ZO",
-            "18",
-        ),
-        (
-            "loc 3",
-            "20",
-            "IB",
-        ),
-        (
-            "loc 4",
-            "ZO",
-            "IB",
-        ),
-    ]
-
-    cast_to_int_expected_rows = [
-        (
-            "loc 1",
-            20,
-            18,
-        ),
-    ]
-
-    cast_to_int_errors_expected_rows = [
-        (
-            "loc 1",
-            20,
-            18,
-        ),
-        (
-            "loc 2",
-            None,
-            18,
-        ),
-        (
-            "loc 3",
-            20,
-            None,
-        ),
-        (
-            "loc 4",
-            None,
-            None,
-        ),
-    ]
-
     small_location_rows = [
         (
             "loc-1",
@@ -2357,6 +2295,68 @@ class CleaningUtilsData:
         ("loc 3", "20220205", "2022", "02", "05"),
         ("loc 5", "20220301", "2022", "03", "01"),
         ("loc 6", "20220402", "2022", "04", "02"),
+    ]
+
+    cast_to_int_rows = [
+        (
+            "loc 1",
+            "20",
+            "18",
+        ),
+    ]
+
+    cast_to_int_errors_rows = [
+        (
+            "loc 1",
+            "20",
+            "18",
+        ),
+        (
+            "loc 2",
+            "ZO",
+            "18",
+        ),
+        (
+            "loc 3",
+            "20",
+            "IB",
+        ),
+        (
+            "loc 4",
+            "ZO",
+            "IB",
+        ),
+    ]
+
+    cast_to_int_expected_rows = [
+        (
+            "loc 1",
+            20,
+            18,
+        ),
+    ]
+
+    cast_to_int_errors_expected_rows = [
+        (
+            "loc 1",
+            20,
+            18,
+        ),
+        (
+            "loc 2",
+            None,
+            18,
+        ),
+        (
+            "loc 3",
+            20,
+            None,
+        ),
+        (
+            "loc 4",
+            None,
+            None,
+        ),
     ]
