+++ tests/test_file_schemas.py
@@ -134,22 +134,6 @@ class ASCWDSWorkplaceSchemas:
         ]
     )
 
-    cast_to_int_schema = StructType(
-        [
-            StructField(AWP.location_id, StringType(), True),
-            StructField(AWP.total_staff, StringType(), True),
-            StructField(AWP.worker_records, StringType(), True),
-        ]
-    )
-
-    cast_to_int_expected_schema = StructType(
-        [
-            StructField(AWP.location_id, StringType(), True),
-            StructField(AWP.total_staff, IntegerType(), True),
-            StructField(AWP.worker_records, IntegerType(), True),
-        ]
-    )
-
     location_schema = StructType(
         [
             StructField(AWP.location_id, StringType(), True),
@@ -1193,6 +1177,22 @@ class CleaningUtilsSchemas:
             StructField(Keys.year, StringType(), True),
             StructField(Keys.month, StringType(), True),
             StructField(Keys.day, StringType(), True),
+        ]
+    )
+
+    cast_to_int_schema = StructType(
+        [
+            StructField(AWP.location_id, StringType(), True),
+            StructField(AWP.total_staff, StringType(), True),
+            StructField(AWP.worker_records, StringType(), True),
+        ]
+    )
+
+    cast_to_int_expected_schema = StructType(
+        [
+            StructField(AWP.location_id, StringType(), True),
+            StructField(AWP.total_staff, IntegerType(), True),
+            StructField(AWP.worker_records, IntegerType(), True),
         ]
     )
