+++ tests/unit/test_cleaning_utils.py
@@ -15,6 +15,10 @@ from tests.test_file_data import CleaningUtilsData as Data
 from utils.column_names.raw_data_files.ascwds_worker_columns import (
     AscwdsWorkerColumns as AWK,
 )
+from utils.column_names.raw_data_files.ascwds_workplace_columns import (
+    PartitionKeys,
+    AscwdsWorkplaceColumns as AWP,
+)
 from utils.column_names.cleaned_data_files.ascwds_workplace_cleaned import (
     AscwdsWorkplaceCleanedColumns as AWPClean,
 )
@@ -595,3 +599,43 @@ class ReduceDatasetToEarliestFilePerMonthTests(unittest.TestCase):
             returned_df.sort(IndCQC.location_id).collect(),
             expected_df.collect(),
         )
+
+
+class CastToIntTests(unittest.TestCase):
+    def setUp(self) -> None:
+        self.spark = utils.get_spark()
+        self.filled_posts_columns = [AWP.total_staff, AWP.worker_records]
+
+    def test_cast_to_int_returns_strings_formatted_as_ints_to_ints(self):
+        cast_to_int_df = self.spark.createDataFrame(
+            Data.cast_to_int_rows, Schemas.cast_to_int_schema
+        )
+        cast_to_int_expected_df = self.spark.createDataFrame(
+            Data.cast_to_int_expected_rows, Schemas.cast_to_int_expected_schema
+        )
+
+        returned_df = job.cast_to_int(cast_to_int_df, self.filled_posts_columns)
+
+        returned_data = returned_df.sort(AWP.location_id).collect()
+        expected_data = cast_to_int_expected_df.sort(AWP.location_id).collect()
+
+        self.assertEqual(expected_data, returned_data)
+
+    def test_cast_to_int_returns_strings_not_formatted_as_ints_as_none(self):
+        cast_to_int_with_errors_df = self.spark.createDataFrame(
+            Data.cast_to_int_errors_rows, Schemas.cast_to_int_schema
+        )
+        cast_to_int_with_errors_expected_df = self.spark.createDataFrame(
+            Data.cast_to_int_errors_expected_rows, Schemas.cast_to_int_expected_schema
+        )
+
+        returned_df = job.cast_to_int(
+            cast_to_int_with_errors_df, self.filled_posts_columns
+        )
+
+        returned_data = returned_df.sort(AWP.location_id).collect()
+        expected_data = cast_to_int_with_errors_expected_df.sort(
+            AWP.location_id
+        ).collect()
+
+        self.assertEqual(expected_data, returned_data)
